fixed movement and added support for entities

+++ src/main/java/com/saulms/cavernwarrior/Camera.java
@@ -2,26 +2,26 @@ package com.saulms.cavernwarrior;
 
 public class Camera {
 
-    private double x, y;
+    private int x, y;
 
-    public Camera(double x, double y) {
+    public Camera(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public double getX() {
+    public int getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(int x) {
         this.x = x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(int y) {
         this.y = y;
     }
